+++ evm/vm/forks/sharding/vmc_utils.py
@@ -5,12 +5,12 @@ from evm.utils.hexadecimal import (
     decode_hex,
 )
 
-from evm.chains.sharding.mainchain_handler.config import GASPRICE
+from evm.vm.forks.sharding.config import GASPRICE
 
 
 def get_vmc_json():
     mydir = os.path.dirname(__file__)
-    vmc_path = os.path.join(mydir, '../contracts/validator_manager.json')
+    vmc_path = os.path.join(mydir, 'contracts/validator_manager.json')
     vmc_json_str = open(vmc_path).read()
     return json.loads(vmc_json_str)
